requirements.txt
redis
fastapi
uvicorn[standard]
requests
testing.redis